Add `pals_to_impactx.py`
Mirror MAD-X Logic Location
